pyproject.toml
[project]
name = "sports-edge"
version = "0.1.0"
description = "AI Probability + Edge Engine for Sportsbook MVP"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ingest", "edge", "features", "models", "backtest", "report", "scripts", "tests"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
